Numbering schema edit
Added illustration number references as parenthetical guides, per the author. Example: Type "knowledge" in the filter navigator (1).

The step numbering, while in order, is a bit out of whack. It's not fatal -- the steps appear in preview in the proper order, it's just that there are skips here and there 1,2,3,6,7,8....I don't have time today to do any rewiring, unfortunately, but will ask Dave S to check it out.